Make command pending errors explicit in `enable` and `dither_enable`
The enable function takes some time for the laser to get stabilized at the right
frequency. This is likely dependent on each individual laser.

There is a way to check the pending bit with nop but this is not implemented
yet. Once it is we may implement a way to wait for pending operations to
complete before progressing. I am not sure if the pending bit stays on for long
enough though.

For now the user will have to handle this explicitly. This will be added to the
primary examples of laser user in the README. This choice is on purpose so that
users can specifically define the wait time that they are comfortable with.

rewriting readme to show explicit error handling in enable

removing CPException handling for dither enable

Again it is important that the user handle these errors themselves. Especially
for experiments it is important that the program does not continue while the
laser is still warming up especially for experiments that rely on very precise
frequency settings such as interacting with a crystal or something like that.